Move Deterministic losses and expousre run into compuation class
