--- oasislmf/computation/run/fm_test.py
@@ -1,16 +1,0 @@
-#__all__ = [
-#    '_'
-#]
-#
-#
-#class _(ComputationStep):
-#    """
-#    - desc here -   
-#    """
-#    step_params = [ {'name': 'oed_location_csv', 'is_path': True, 'pre_exist': True,
-#                    'help': 'Source location CSV file path'},
-#                   ]
-#
-#
-#    def run(self):
-#        return None
